+++ build_web_compilers/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 3.1.0-dev
 
-- Add the `generate-debug-symbols` option for the `build_web_compilers:ddc`
+- Add the `emit-debug-symbols` option for the `build_web_compilers:ddc`
   builder. The debug symbols output is used by the debugger in webdev for
   variable inspection. This setting is disabled by default but can be
   enabled by setting it to `true` globally:
@@ -9,7 +9,7 @@
 global_options:
   build_web_compilers:ddc:
     options:
-      generate-debug-symbols: true
+      emit-debug-symbols: true
 ```